Authz: Move extension proto up a layer
