+++ pkg/services/accesscontrol/dualwrite/collectors.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/grafana/grafana/pkg/infra/db"
 	"github.com/grafana/grafana/pkg/services/authz/zanzana"
-	authzextv1 "github.com/grafana/grafana/pkg/services/authz/zanzana/proto/v1"
+	authzextv1 "github.com/grafana/grafana/pkg/services/authz/proto/v1"
 )
 
 func teamMembershipCollector(store db.DB) legacyTupleCollector {
