+++ pkg/services/accesscontrol/dualwrite/resource_reconciler.go
@@ -7,8 +7,8 @@ import (
 	"github.com/grafana/authlib/claims"
 	openfgav1 "github.com/openfga/api/proto/openfga/v1"
 
+	authzextv1 "github.com/grafana/grafana/pkg/services/authz/proto/v1"
 	"github.com/grafana/grafana/pkg/services/authz/zanzana"
-	authzextv1 "github.com/grafana/grafana/pkg/services/authz/zanzana/proto/v1"
 )
 
 // legacyTupleCollector collects tuples groupd by object and tupleKey
