+++ pkg/services/authz/proto/v1/buf.gen.yaml
@@ -1,10 +1,10 @@
 version: v1
 plugins:
   - plugin: go
-    out: pkg/services/authz/zanzana/proto/v1
+    out: pkg/services/authz/proto/v1
     opt: paths=source_relative
   - plugin: go-grpc
-    out: pkg/services/authz/zanzana/proto/v1
+    out: pkg/services/authz/proto/v1
     opt:
       - paths=source_relative
       - require_unimplemented_servers=false
