+++ pkg/services/authz/zanzana/client/noop.go
@@ -6,7 +6,7 @@ import (
 	"github.com/grafana/authlib/authz"
 	"github.com/grafana/authlib/claims"
 
-	authzextv1 "github.com/grafana/grafana/pkg/services/authz/zanzana/proto/v1"
+	authzextv1 "github.com/grafana/grafana/pkg/services/authz/proto/v1"
 )
 
 var _ authz.AccessClient = (*NoopClient)(nil)
